feat: moved common function to commonutil class
