+++ cmd/collectors/rest/plugins/snapmirror/snapmirror.go
@@ -167,7 +167,7 @@ func (my *SnapMirror) updateSMLabels(data *matrix.Matrix) {
 		collectors.UpdateProtectedFields(instance)
 
 		// Update lag time based on checks
-		my.updateLagTime(instance, lastTransferSizeMetric, lagTimeMetric)
+		collectors.UpdateLagTime(instance, lastTransferSizeMetric, lagTimeMetric, my.Logger)
 	}
 
 	// handle CG relationships
@@ -211,26 +211,6 @@ func (my *SnapMirror) handleCGRelationships(data *matrix.Matrix, keys []string) 
 				cgVolumeInstance.SetLabel("source_volume", sourceVol)
 				cgVolumeInstance.SetLabel("destination_volume", destinationVol)
 			}
-		}
-	}
-
-}
-
-func (my *SnapMirror) updateLagTime(instance *matrix.Instance, lastTransferSize *matrix.Metric, lagTime *matrix.Metric) {
-	healthy := instance.GetLabel("healthy")
-	schedule := instance.GetLabel("schedule")
-	lastError := instance.GetLabel("last_transfer_error")
-	relationshipID := instance.GetLabel("relationship_id")
-
-	// If SM relationship is healthy, has a schedule, last_transfer_error is empty, and last_transfer_bytes is 0, Then we are setting lag_time to 0
-	// Otherwise, report the lag_time which ONTAP has originally reported.
-	if lastBytes, ok := lastTransferSize.GetValueFloat64(instance); ok {
-		if healthy == "true" && schedule != "" && lastError == "" && lastBytes == 0 {
-			lag, _ := lagTime.GetValueFloat64(instance)
-			if err := lagTime.SetValueFloat64(instance, 0); err != nil {
-				my.Logger.Error().Err(err).Str("metric", lagTime.GetName()).Msg("Unable to set value on metric")
-			}
-			my.Logger.Debug().Msgf("lagTime value set from %f to 0 for %s. Healthy: %s, Schedule: %s, LastBytes: %f, LastError:%s", lag, relationshipID, healthy, schedule, lastBytes, lastError)
 		}
 	}
