+++ cmd/collectors/zapi/plugins/snapmirror/snapmirror.go
@@ -11,7 +11,6 @@ import (
 	"github.com/netapp/harvest/v2/pkg/conf"
 	"github.com/netapp/harvest/v2/pkg/dict"
 	"github.com/netapp/harvest/v2/pkg/errs"
-	"github.com/netapp/harvest/v2/pkg/logging"
 	"github.com/netapp/harvest/v2/pkg/matrix"
 	"github.com/netapp/harvest/v2/pkg/tree/node"
 	"regexp"
@@ -117,7 +116,7 @@ func (my *SnapMirror) Run(dataMap map[string]*matrix.Matrix) ([]*matrix.Matrix, 
 			collectors.UpdateProtectedFields(instance)
 
 			// Update lag time based on checks
-			UpdateLagTime(instance, lastTransferSizeMetric, lagTimeMetric, my.Logger)
+			collectors.UpdateLagTime(instance, lastTransferSizeMetric, lagTimeMetric, my.Logger)
 		} else {
 			// 7 Mode
 			// source / destination nodes can be something like:
@@ -225,24 +224,4 @@ func (my *SnapMirror) getSVMPeerData(cluster string) error {
 		my.svmPeerDataMap[localSvmName] = Peer{svm: actualSvmName, cluster: peerClusterName}
 	}
 	return nil
-}
-
-func UpdateLagTime(instance *matrix.Instance, lastTransferSize *matrix.Metric, lagTime *matrix.Metric, logger *logging.Logger) {
-	healthy := instance.GetLabel("healthy")
-	schedule := instance.GetLabel("schedule")
-	lastError := instance.GetLabel("last_transfer_error")
-	relationshipID := instance.GetLabel("relationship_id")
-
-	// If SM relationship is healthy, has a schedule, last_transfer_error is empty, and last_transfer_bytes is 0, Then we are setting lag_time to 0
-	// Otherwise, report the lag_time which ONTAP has originally reported.
-	if lastBytes, ok := lastTransferSize.GetValueFloat64(instance); ok {
-		if healthy == "true" && schedule != "" && lastError == "" && lastBytes == 0 {
-			lag, _ := lagTime.GetValueFloat64(instance)
-			if err := lagTime.SetValueFloat64(instance, 0); err != nil {
-				logger.Error().Err(err).Str("metric", lagTime.GetName()).Msg("Unable to set value on metric")
-			}
-			logger.Debug().Msgf("lagTime value set from %f to 0 for %s. Healthy: %s, Schedule: %s, LastBytes: %f, LastError:%s", lag, relationshipID, healthy, schedule, lastBytes, lastError)
-		}
-	}
-
 }
