+++ cmd/collectors/zapi/plugins/snapmirror/snapmirror_test.go
@@ -2,7 +2,6 @@ package snapmirror
 
 import (
 	"github.com/netapp/harvest/v2/cmd/collectors"
-	"github.com/netapp/harvest/v2/pkg/logging"
 	"github.com/netapp/harvest/v2/pkg/matrix"
 	"testing"
 )
@@ -211,65 +210,4 @@ func testWithNoPolicyType(t *testing.T, instance *matrix.Instance) {
 	} else {
 		t.Errorf("Labels derived_relationship_type= %s, expected: extended_data_protection", instance.GetLabel("derived_relationship_type"))
 	}
-}
-
-func TestRuleSimpleLatencyAggregation(t *testing.T) {
-	type test struct {
-		name              string
-		instance          string
-		healthy           string
-		schedule          string
-		lastTransferError string
-		relationshipID    string
-		lastBytesValue    float64
-		lagTimeValue      float64
-		want              float64
-	}
-	tests := []test{
-		{name: "TestBytesNonZero", instance: "InstanceA", healthy: "true", schedule: "5min", lastTransferError: "", relationshipID: "4885136b-4c31-11ec-95a6-00a09865cd13", lastBytesValue: 3479, lagTimeValue: 172652, want: 172652},
-		{name: "TestLastErrorNonEmpty", instance: "InstanceB", healthy: "true", schedule: "my_daily", lastTransferError: "The specified Snapshot copy is older than the base Snapshot copy on source volume", relationshipID: "7c4d92e1-6828-11ea-893b-00a09865cd13", lastBytesValue: 0, lagTimeValue: 172223, want: 172223},
-		{name: "TestScheduleEmpty", instance: "InstanceC", healthy: "true", schedule: "", lastTransferError: "", relationshipID: "62422099-9c61-11e8-b2ff-00a09865fe59", lastBytesValue: 0, lagTimeValue: 738, want: 738},
-		{name: "TestNonHealthy", instance: "InstanceD", healthy: "false", schedule: "daily", lastTransferError: "", relationshipID: "31421f02-7703-11e9-b59e-00a09865cd13", lastBytesValue: 0, lagTimeValue: 892, want: 892},
-		{name: "TestLagTimeChangedTo0", instance: "InstanceE", healthy: "true", schedule: "hourly", lastTransferError: "", relationshipID: "b4da7644-76e0-11e9-b59e-00a09865cd13", lastBytesValue: 0, lagTimeValue: 2736, want: 0},
-	}
-
-	data := matrix.New("SnapMirrorTest", "SnapMirror", "SnapMirror")
-	lastTransferSizeMetric, err := data.NewMetricUint8("last_transfer_size")
-	if err != nil {
-		panic(err)
-	}
-	lagTimeMetric, err := data.NewMetricUint8("lag_time")
-	if err != nil {
-		panic(err)
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			instance := populateInstance(data, tt.instance, tt.healthy, tt.schedule, tt.lastTransferError, tt.relationshipID, lastTransferSizeMetric, tt.lastBytesValue, lagTimeMetric, tt.lagTimeValue)
-			UpdateLagTime(instance, lastTransferSizeMetric, lagTimeMetric, logging.Get())
-			actualValue, _ := lagTimeMetric.GetValueFloat64(instance)
-			if actualValue != tt.want {
-				t.Errorf("expected %f got %f", tt.want, actualValue)
-			}
-		})
-	}
-}
-
-func populateInstance(data *matrix.Matrix, instanceName, healthy, schedule, lastError, relationshipID string, lastTransferSizeMetric *matrix.Metric, bytesData float64, lagTimeMetric *matrix.Metric, lagTime float64) *matrix.Instance {
-	instance, err := data.NewInstance(instanceName)
-	if err != nil {
-		panic(err)
-	}
-	instance.SetLabel("healthy", healthy)
-	instance.SetLabel("schedule", schedule)
-	instance.SetLabel("last_transfer_error", lastError)
-	instance.SetLabel("relationship_id", relationshipID)
-
-	if err = lastTransferSizeMetric.SetValueFloat64(instance, bytesData); err != nil {
-		panic(err)
-	}
-	if err = lagTimeMetric.SetValueFloat64(instance, lagTime); err != nil {
-		panic(err)
-	}
-	return instance
 }
